Loader now Load the core first then plugins. Added LobbyServer for hosting. Added ACS.dll to make people see whats inside. (Also that it needs to copy that) Add Config option to lobby,ip,port, maxplayer Changed from AppData path to Game folder. Changed fully NetBirdApi to use LobbyServer Removed watermark, credits, etc Removed lang option, defaulting to en-US (might come back to use others) Fixing MultiplayerCreatureItem.IsFrozen Removed custom Ending credits

+++ Subnautica.Core/Subnautica.Network/Models/Client/JoiningServerArgs.cs
@@ -136,7 +136,7 @@ namespace Subnautica.Network.Models.Client
         public byte PlayerId { get; set; }
 
         [Key(42)]
-        public byte MaxPlayerCount { get; set; }
+        public int MaxPlayerCount { get; set; }
 
         [Key(43)]
         public Dictionary<string, string> SeaTruckConnections { get; set; }
